Refactor : detail directory
