+++ src/App.jsx
@@ -6,13 +6,13 @@ import Footer from './layout/Footer';
 import MainPage from './main';
 import CentralClub from './menu/central_club/central_club';
 import SmallClub from './menu/small_club/small_club';
-import DetailPage from './menu/small_club/detail_page/index';
+import DetailPage from './menu/detail/index';
 import BranchCentral from './menu/central_club/branch/branchCentral';
 import LoginPage from './menu/login';
 import KakaoRedirection from './menu/login/component/kakaoRedirection';
 import BranchSmall from './menu/small_club/branch/branch_small';
 import BranchHashTag from './components/hashtag/component/branchHashtag';
-import ReviewWrite from './menu/small_club/detail_page/review_write';
+import ReviewWrite from './menu/detail/review_write';
 import Search from './menu/search/search';
 
 function App() {
@@ -36,8 +36,8 @@ function App() {
                         <Route path="/menu/central_club/central_club" element={<CentralClub />} />
                         {/* <Route path="/menu/central_club/detail_page/review_write" element={<ReviewWrite />} /> */}
                         <Route path="/menu/small_club/small_club" element={<SmallClub />} />
-                        <Route path="/menu/small_club/detail_page/:clubId" element={<DetailPage />} />
-                        <Route path="/menu/small_club/detail_page/review_write" element={<ReviewWrite />} />
+                        <Route path="/menu/detail/:clubId" element={<DetailPage />} />
+                        <Route path="/menu/detail/review_write" element={<ReviewWrite />} />
                         <Route path="/menu/central_club/branch/branchCentral" element={<BranchCentral />} />
                         <Route path="/menu/login" element={<LoginPage />} />
                         <Route path="/v1/auths/oauth/kakao" element={<KakaoRedirection />} />
@@ -56,8 +56,8 @@ function App() {
                         <Route path="/" element={<MainPage />} />
                         <Route path="/menu/central_club/central_club" element={<CentralClub />} />
                         <Route path="/menu/small_club/small_club" element={<SmallClub />} />
-                        <Route path="/menu/small_club/detail_page/:clubId" element={<DetailPage />} />
-                        <Route path="/menu/small_club/detail_page/review_write" element={<ReviewWrite />} />
+                        <Route path="/menu/detail/:clubId" element={<DetailPage />} />
+                        <Route path="/menu/detail/review_write" element={<ReviewWrite />} />
                         <Route path="/menu/central_club/branch/branchCentral" element={<BranchCentral />} />
                         <Route path="/menu/login" element={<LoginPage />} />
                         <Route path="/v1/auths/oauth/kakao" element={<KakaoRedirection />} />
