+++ src/menu/detail/_component/ReviewStatics.jsx
