+++ src/menu/detail/_component/reviewStats.css
