+++ src/menu/detail/_component/reviewTab.css
