+++ src/menu/detail/component/clubDetail.jsx
