+++ src/menu/detail/detailPage.css
