+++ src/menu/detail/index.jsx
@@ -1,5 +1,4 @@
 import { useState, useEffect } from 'react';
-import { useParams } from 'react-router-dom';
 import axios from 'axios';
 import IntroductionTab from './_component/IntroductionTab';
 import ReviewTab from './_component/ReviewTab';
